+++ react/components/MenuContext.tsx
@@ -1,6 +1,7 @@
 import { createContext } from 'react'
 
 const MenuContext = createContext<MenuContextValue>({
+  experimentalOptimizeRendering: false,
   hasTitle: false,
   orientation: 'horizontal',
   textType: 't-body',
@@ -10,6 +11,7 @@ interface MenuContextValue {
   hasTitle: boolean
   orientation: 'horizontal' | 'vertical'
   textType: string
+  experimentalOptimizeRendering?: boolean
 }
 
 export default MenuContext