+++ react/__mocks__/vtex.css-handles.tsx
@@ -1,8 +1,13 @@
-export const useCssHandles = (cssHandles: string[]) => {
+export const useCssHandles = (
+  cssHandles: string[],
+  _?: Record<string, any>
+) => {
   const handles: any = {}
   cssHandles.forEach(handle => {
     handles[handle] = handle
   })
 
-  return handles
+  const withModifiers = (baseHandle: string, __: any) => baseHandle
+
+  return { handles, withModifiers }
 }