+++ react/typings/vtex.css-handles.d.ts
@@ -1,5 +1,0 @@
-declare module 'vtex.css-handles' {
-  export const useCssHandles
-  export const withCssHandles
-  export const applyModifiers
-}