+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Add prop 'isOpenOnMount' to allow opening submenu on mount
 ## [2.27.0] - 2020-11-05
 ### Changed
 - Close submenus whenever the URL changes.