+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- I18n Ro and Jp.
+
+### Fixed
+
+- Crowdin configuration file.
+
 ## [2.28.0] - 2021-01-04
 ### Added
 - New `submenuColumn` CSS handle.